BEST is a leader in IPC training and Certification
IPC Training –What is it and Why do I Need It?
The IPC, the trade organizations for the PCB and electronics, has developed a series of both inspection criteria and hand skills courses related to PCB s and their assembly. Industry consensus within the IPC has driven the desire for a training program. For the IPC training programs involving inspection criteria the classroom teaching involves being able understand the specification and being able to look up the material. For the hands on IPC solder training program skills in the subject matter like soldering, rework and repair of printed circuit boards need to be demonstrated along with proficiency in the specification.
These IPC training courses are based on the industry specifications assembled via the IPC trade organization based on member input and cover a variety of topics. These along with IPC member input drive the training programs. Specifications are created for the given topic through the input of members at committee meetings based on volunteer support. The supporting members are training centers, consultants to the industry and end users. These working groups, chaired by an industry member and organized and guided by IPC staff members, reach agreement on the specification. After a final vote by industry the document is published and subsequently translated into a variety of different languages. These documents are on the basis for the IPC training program. A separate training committee is then established and drives the training program content. After a few trial classes and one or more rounds of IPC training program, the IPC training program is then released.
The IPC training programs are administered by the IPC, the industry trade organization. Administration of the IPC training program entails keeping track of instructor credentials, verifying the training center’s competence as well as printing and distributing the materials and maintaining the web site. In addition to maintaining the web site, the IPC administers the testing portal.
The IPC certification training programs are taught by subject matter experts from industry using IPC coordinated and industry-approved materials. IPC credentialed, authorized and licensed training centres make sure that their instructors are vetted requiring a minimum amount of classroom and industry experience.  In turn, the IPC training centers make sure that the newly-licensed instructors going back to their locations are able to teach the materials for the class to their own associates as part of the IPC training.
An IPC certification, which is the result of a successful outcome of IPC training, is governed by a set of agreed guidelines and rules. These rules are setup in a living document called the “The IPC Governing Policy and Procedures for Training”. Members of this committee drive policies and requirements for IPC training.
There are several drivers for the need for IPC certification training. One of the drivers is the desire and requirement by space, governmental bodies as well as international companies from having their products built to and by people knowledgeable in industry standards. These standards are accepted worldwide and allow the owner of the design to have choices in where their products should be built. The IPC certification in many cases faces a requirement for those who will be building, assembling and testing their products.  On the flip side, the industry in then driven by this requirement and demand to have certified and trained personnel in order to participate in these contracts. A second driver for companies to have their employees participate in the IPC certification program in order to meet the ISO and QS quality system requirements for having a 3rd-part certified training program. Each of the IPC certification training programs matches these requirements as they are linked back to the ANSI standard.
The IPC certification program is constantly evolving its requirements. The local licensed training centers with their subject matter, experts are there to help users and companies to navigate the requirements. In addition, the IPC training centers are able to teach the standards and hand skills as their staff teaches full time and has the commensurate experience to bring the training to life. Look to your local IPC training center to bring these requirements and standards to life.
 